# platform-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hyperfleet Platform API
  description: Platform API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /clusters:
    get:
      summary: List user's clusters
      description: Retrieve clusters owned by the authenticated user
      operationId: listUserClusters
      tags:
        - Clusters
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of clusters to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of clusters to skip (pagination)
      responses:
        "200":
          description: List of user's clusters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      summary: Create cluster
      description: |
        Create a new cluster for the authenticated user.

        The API automatically enriches the cluster spec with:
        - hostedCluster.issuerURL: Auto-populated from the regional OIDC CloudFront distribution
        - hostedCluster.release.image: Defaulted if not provided
        - hostedCluster.networking: Default CIDR ranges applied if not provided

        Possible error codes:
        - CLUSTERS-MGMT-CREATE-001: Invalid request body
        - CLUSTERS-MGMT-CREATE-002: Invalid cluster spec
        - CLUSTERS-MGMT-CREATE-003: Failed to create cluster
        - CLUSTERS-MGMT-CREATE-004: Failed to validate cluster name
        - CLUSTERS-MGMT-CREATE-005: Duplicate cluster name
      operationId: createUserCluster
      tags:
        - Clusters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterCreateRequest"
      responses:
        "201":
          description: Cluster created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /clusters/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Cluster ID

    get:
      summary: Get cluster details
      description: Retrieve cluster details (user must own the cluster)
      operationId: getUserCluster
      tags:
        - Clusters
      responses:
        "200":
          description: Cluster details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      summary: Update cluster
      description: Update cluster spec (user must own the cluster)
      operationId: patchUserCluster
      tags:
        - Clusters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterUpdateRequest"
      responses:
        "200":
          description: Cluster updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update cluster
      description: Update cluster spec (user must own the cluster). Same behavior as PATCH.
      operationId: updateUserCluster
      tags:
        - Clusters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterUpdateRequest"
      responses:
        "200":
          description: Cluster updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete cluster
      description: Delete cluster (user must own the cluster)
      operationId: deleteUserCluster
      tags:
        - Clusters
      responses:
        "202":
          description: Cluster deletion initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cluster_id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      required:
        - kind
        - code
        - reason
      properties:
        kind:
          type: string
          description: Resource kind
          example: Error
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
...
    ClusterSpec:
      type: object
      description: |
        Cluster specification following the hyperfleet-operator v1alpha1.ClusterSpec
        type. Contains a `creatorARN` and a nested `hostedCluster` field that follows
        the HyperShift v1beta1 HostedClusterSpec schema.
      properties:
        creatorARN:
          type: string
          description: ARN of the user who created this cluster (auto-populated by the API)
        hostedCluster:
          type: object
          description: |
            HyperShift v1beta1 HostedClusterSpec. Key fields include:
            - platform: Cloud provider configuration (type, aws)
            - networking: Cluster/service/machine network CIDRs
            - release: OpenShift release image
            - issuerURL: OIDC issuer URL (auto-populated)
          additionalProperties: true
...
    Cluster:
      type: object
      description: A user cluster resource
      required:
        - id
        - name
        - target_project_id
        - created_by
        - generation
        - resource_version
        - spec
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the cluster
        name:
          type: string
          description: Cluster name
        target_project_id:
          type: string
          description: Target project ID (AWS account ID)
        created_by:
          type: string
          description: ARN of the user who created this cluster
        oidc_issuer_url:
          type: string
          description: OIDC issuer URL for this cluster
          example: https://doku78iof5s87.cloudfront.net/cluster-123
        generation:
          type: integer
          format: int64
          description: Generation number for optimistic concurrency control
        resource_version:
          type: string
          description: Resource version for optimistic concurrency control
        spec:
          $ref: "#/components/schemas/ClusterSpec"
        status:
          $ref: "#/components/schemas/ClusterStatusInfo"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    ClusterCreateRequest:
      type: object
      description: |
        Request body for creating a cluster.

        The API automatically enriches the cluster spec with:
        - hostedCluster.issuerURL: Auto-populated from the regional OIDC CloudFront distribution
        - hostedCluster.release.image: Defaulted if not provided
        - hostedCluster.networking: Default CIDR ranges applied if not provided
        - creatorARN: Set from the caller's IAM identity
      required:
        - name
        - spec
      properties:
        name:
          type: string
          pattern: "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
          minLength: 3
          maxLength: 63
          description: DNS-compatible cluster name
        target_project_id:
          type: string
          description: Target project ID
        spec:
          $ref: "#/components/schemas/ClusterSpec"
    ClusterUpdateRequest:
      type: object
      description: Request body for updating a cluster
      required:
        - spec
      properties:
        spec:
          $ref: "#/components/schemas/ClusterSpec"
    ClusterStatusInfo:
      type: object
      description: Aggregated status for clusters
      properties:
        observedGeneration:
          type: integer
          format: int64
          description: Last generation processed by controllers
        conditions:
          type: array
          description: Status conditions
          items:
            $ref: "#/components/schemas/Condition"
        phase:
          type: string
          enum: [WaitingForPlacement, Provisioning, Ready, Deleting]
          description: High-level cluster phase
        controlPlaneEndpoint:
          $ref: "#/components/schemas/APIEndpoint"
        version:
          type: string
          description: Current cluster version
        placementRef:
          $ref: "#/components/schemas/PlacementReference"
        message:
          type: string
          description: Human-readable status message
        reason:
          type: string
          description: Machine-readable reason code
        lastUpdateTime:
          type: string
          format: date-time
          description: When status was last calculated
    APIEndpoint:
      type: object
      description: Control plane API endpoint
      properties:
        host:
          type: string
        port:
          type: integer
          format: int32
    PlacementReference:
      type: object
      description: Management cluster assignment
      properties:
        name:
          type: string
          description: Placement name
        managementCluster:
          type: string
          description: Management cluster name
    Condition:
      type: object
      description: Status condition
      properties:
        type:
          type: string
          description: Condition type
        status:
          type: string
          enum: ["True", "False", Unknown]
          description: Condition status
        lastTransitionTime:
          type: string
          format: date-time
          description: Last time the condition transitioned
        reason:
          type: string
          description: Machine-readable reason code
        message:
          type: string
          description: Human-readable message
...
    ClusterList:
      type: object
      description: Paginated list of clusters
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Cluster"
        total:
          type: integer
          description: Total number of clusters
        limit:
          type: integer
          description: Number of items per page
        offset:
          type: integer
          description: Number of items skipped

    # NodePool Schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"